Move analysis-related stuff -> ir/analysis/...
