+++ devito/ir/__init__.py
@@ -1,3 +1,4 @@
-from devito.ir.dfg import *  # noqa
+from devito.ir.analysis import *  # noqa
 from devito.ir.clusters import *  # noqa
+from devito.ir.dfg import *  # noqa
 from devito.ir.iet import *  # noqa
