+++ devito/ir/analysis/__init__.py
@@ -1,0 +1,1 @@
+from devito.ir.analysis.flow import *  # noqa
